Shopping_Site/Shopping_Site/requirements.txt
asgiref==3.11.0
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-api-logger==1.1.21
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.2
webencodings==0.5.1
